[refactor] lotto, winningLotto 테스트 분리
